# Conflicts:
#	src/main/java/com/green/muziuniv_be_notuser/shared/course/CourseRepository.java

+++ src/main/java/com/green/muziuniv_be_notuser/professor/attendance/model/AttendanceRes.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class AttendanceRes {
     private String attendDate;
-    private Integer enrollmentId;
+    private Long enrollmentId;
     private String status;
     private String note;
 }
